resumecreation/requirements.txt
Django==4.2.15
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.7
PyJWT==2.9.0
python-dotenv==1.0.1
sqlparse==0.5.1
pytz==2024.1
asgiref==3.8.1
decorator==5.1.1
debugpy==1.8.5
h11==0.14.0
sniffio==1.3.1
wsproto==1.2.0
rich==13.7.1